LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/email-set-value-expected.txt



FAIL setValue(sanitizedValue) is reflected in visible text field content assert_equals: value is sanitized expected "foo@bar" but got ""
